+++ src/senaite/sync/browser/views.py
@@ -55,36 +55,31 @@ class Sync(BrowserView):
 
         # Handle form submit
         form = self.request.form
-        fetchform = form.get("fetchform", False)
-        dataform = form.get("dataform", False)
-        if not any([fetchform, dataform]):
+
+        if not form.get("dataform", False):
             return self.template()
+
+        domain_name = form.get("domain_name", None)
+
+        # Handle "Clear this Storage" action
+        if form.get("clear_storage", False):
+            del self.storage[domain_name]
+            delete_soup(self.portal, domain_name)
+            message = _("Cleared Storage {}".format(domain_name))
+            self.add_status_message(message, "info")
+            return self.template()
+
+        # Get the necessary data for the domain
+        storage = self.get_storage(domain_name)
+        credentials = storage["credentials"]
+        config = storage["configuration"]
 
         # Handle "Import" action
         if form.get("import", False):
-            domain_name = form.get("domain_name", None)
-            # initialize the session
-            storage = self.get_storage(domain_name)
-            url = storage["credentials"]["url"]
-            username = storage["credentials"]["username"]
-            password = storage["credentials"]["password"]
-            content_types = storage["configuration"].get("content_types", None)
-            data = {
-                "url": url,
-                "domain_name": domain_name,
-                "ac_name": username,
-                "ac_password": password,
-                "content_types": content_types,
-            }
-            step = ImportStep(data)
-            step.run()
-            return self.template()
+            step = ImportStep(credentials, config)
 
         # Handle "Complement" action
-        if form.get("complement", False):
-            domain_name = form.get("domain_name", None)
-            storage = self.get_storage(domain_name)
-
+        else:
             fetch_time = form.get("mod_date_limit", None) or \
                 storage.get("last_fetch_time", None)
             if not fetch_time:
@@ -100,77 +95,19 @@ class Sync(BrowserView):
                     self.add_status_message(message, "error")
                     return self.template()
 
-            url = storage["credentials"]["url"]
-            username = storage["credentials"]["username"]
-            password = storage["credentials"]["password"]
-            content_types = storage["configuration"].get("content_types", None)
-            data = {
-                "url": url,
-                "domain_name": domain_name,
-                "ac_name": username,
-                "ac_password": password,
-                "fetch_time": fetch_time,
-                "content_types": content_types,
-            }
-            step = ComplementStep(data)
-            step.run()
-            return self.template()
+            step = ComplementStep(credentials, config, fetch_time)
 
-        # Handle "Clear this Storage" action
-        if form.get("clear_storage", False):
-            domain = form.get("domain_name", None)
-            del self.storage[domain]
-            delete_soup(self.portal, domain)
-            message = _("Cleared Storage {}".format(domain))
-            self.add_status_message(message, "info")
-            return self.template()
-
-        # Handle "Fetch" action
-        if form.get("fetch", False):
-
-            url = form.get("url", "")
-            if not url.startswith("http"):
-                url = "http://{}".format(url)
-            domain_name = form.get("domain_name", None)
-            username = form.get("ac_name", None)
-            password = form.get("ac_password", None)
-            # check if all mandatory fields have values
-            if not all([domain_name, url, username, password]):
-                message = _("Please fill in all required fields")
-                self.add_status_message(message, "error")
-                return self.template()
-
-            import_settings = True if form.get("import_settings") == 'on' else False
-            import_users = True if form.get("import_users") == 'on' else False
-            import_registry = True if form.get("import_registry") == 'on' else False
-            content_types = form.get("content_types", None)
-            if content_types is not None:
-                content_types = [t.strip() for t in content_types.split(",")]
-                portal_types = api.get_tool("portal_types")
-                content_types = filter(lambda ct: ct in portal_types,
-                                       content_types)
-
-            data = {
-                "url": url,
-                "domain_name": domain_name,
-                "ac_name": username,
-                "ac_password": password,
-                "content_types": content_types,
-                "import_settings": import_settings,
-                "import_users": import_users,
-                "import_registry": import_registry,
-            }
-
-            fs = FetchStep(data)
-            verified, message = fs.verify()
-            if verified:
-                fs.run()
-                self.add_status_message(message, "info")
-            else:
-                self.add_status_message(message, "error")
-
-        # always render the template
+        step.run()
         return self.template()
+
+    def get_storage_config(self, domain_name, config_name, default = None):
+        """ Get the advanced configuration setting for a given domain
+        :param config_name: advanced configuration section name
+        :param default: default value if configuration value is not set
+        :return:
+        """
+        storage = self.get_storage(domain_name)
+        return storage["configuration"].get(config_name, default)
 
     def add_status_message(self, message, level="info"):
         """Set a portal status message